+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Versioning].
 
 ## [Unreleased]
 
-- _No changes yet_
+- Provide testing utilities. ([#710])
 
 ## [1.6.5] - 2023-01-26
 
@@ -233,6 +233,7 @@ Versioning].
 [#666]: https://github.com/ericcornelissen/shescape/pull/666
 [#675]: https://github.com/ericcornelissen/shescape/pull/675
 [#682]: https://github.com/ericcornelissen/shescape/pull/682
+[#710]: https://github.com/ericcornelissen/shescape/pull/710
 [552e8ea]: https://github.com/ericcornelissen/shescape/commit/552e8eab56861720b1d4e5474fb65741643358f9
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html